feature(authentication):signin and signup Currently, users cannot signup or signin because the feature is yet to be implemented which makes the app unusable.
This commits implements the signin and signup feature
to enable users create an account and login to the app
for them to be able to explore the other functionalities of the app.
This feature was implemented by creating classes 'register' and 'signin'
in the 'UsersController'.
Also, the table structure of the database is adjusted to include a
few other implementation